ajust the 'add new board' to be near the existed boards by add it to <ul> of BoardList components
