+++ src/components/BoardHome/BoardHome.js
@@ -1,12 +1,8 @@
-import React, { useRef } from "react";
+import React from "react";
 import { BoardHomeHeader } from "./BoardHomeHeader";
 import { BoardLists } from "./BoardLists";
-import { useDispatch } from "react-redux";
-import { createNewBoard } from "../../store/actions/boardActions";
 
 export function BoardHome() {
-  const newBoardName = useRef(null);
-  const dispatch = useDispatch();
   return (
     <div id="home">
       <BoardHomeHeader></BoardHomeHeader>
@@ -16,17 +12,6 @@ export function BoardHome() {
       </h2>
       <ul className="board-items">
         <BoardLists></BoardLists>
-        <li className="board-item create-new-board">
-          <textarea
-            className="title form-field-input"
-            placeholder="创建新看板"
-            ref={newBoardName}
-            onBlur={() => {
-              dispatch(createNewBoard({ name: newBoardName.current.value }));
-              newBoardName.current.value = "";
-            }}
-          ></textarea>
-        </li>
       </ul>
     </div>
   );
